refactor(audio): extract AudioDevice and share the plumbing-name prefix
The aggregate-name contract was stringly and split across three files:
ProcessTap and MultiOutputController produced names starting "Fader",
AudioDeviceMonitor filtered on the literal — renaming a producer would
silently leak Fader's own aggregates into the user-facing device list.
One shared constant plus an isFaderPlumbing predicate closes that class
of break.

AudioDevice moves to its own file so upcoming pure-policy code (and the
hostless test bundle) can use it without dragging in the monitor;
symbolName moves UI-side, keeping presentation vocabulary out of Audio/.
